requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Data Processing & ML
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
shap==0.44.0
joblib==1.3.2

# Blockchain & Web3
web3==6.11.3
requests==2.31.0
aiohttp==3.9.1
asyncio-throttle==1.0.2

# AI & NLP
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Rate Limiting & Caching
slowapi==0.1.9
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP Client
httpx==0.25.2

# Environment & Configuration
pydantic-settings==2.1.0